ci: remove workaround for incompatibility between lerna@7 and npm@9 (#398)
Lerna cannot bump versions due to incompatibility of the lock
file, so it was ignored during the version bumping.

This workaround introduced a bug mentioned at https://github.com/ckb-js/kuai/issues/389

Now I've tried to bump versions without ignoring the lock file
and it works, so the workaround is simply removed by this commit.

Ref: https://github.com/lerna/lerna/issues/3418#issuecomment-1337756237